feat(front/withdraw): update forms and create base components
